Save to file functionality; major refactoring of memory representation of a record
Now the program will save automatically to the file set from the output
menu (choice 3 from the start menu).
A major refactoring of the memory representation was made - now this
class has vector with pointers to fields from a base class for fields
instead of listing all the fields. The refactoring made available more
automatically working with these fields