Code refactory
Changes:
Code refactory: preparation for the program to be able to load/save in
multiple file formats (will be made in next version);
Removed (for now) possibility on saving as a txt file.
The program prints on screen the current progress.
Corrections on some info messages the program prints on screen.
Some improvements on program speed.